+++ lms/user.py
@@ -3026,11 +3026,12 @@ def contact_us(**kwargs):
             subject = "Contact us Request – " + getseries("Contact us Request –", 3)
             frappe.db.commit()
 
-            message = "From {name},<br>Email id - {email},<br>Customer id - {cust},<br><br>{mess}".format(
+            message = "{mess}<br><br>From {name},<br>Email id - {email},<br>Mobile number - {phone},<br>Customer id - {cust}".format(
+                mess=data.get("message").strip(),
                 name=user.full_name,
                 email=user.email,
+                phone=lms.__customer().phone,
                 cust=lms.__customer().name,
-                mess=data.get("message").strip(),
             )
 
             frappe.get_doc(